+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Version 1.5.0 (2020-11-06)
 
 * [new] Support configuration of injectable custom error decoders using the `errorDecoder` configuration property.
+* [new] Support configuration of injectable retryers using the `retryer` configuration property (global level or endpoint level).
+* [new] Support configuration of standard retry parameters using the `retry.period` / `retry.maxPeriod` / `retry.maxAttempts` properties (global or endpoint level).
+
 # Version 1.4.0 (2020-08-05)
 
 * [new] Support configuration of injectable endpoint request interceptors using the `interceptors` configuration property.